+++ game-balatro/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
-Future Changes (NOT IMPLEMENTED YET):
+## Future Changes (NOT IMPLEMENTED YET)
 
-- NONE
+- Xbox support - waiting on stable (and easy to install) mod injector for Xbox version
+- Automatic update checks for Lovely-Injector - need to solve technical issues with version resolution (no version in filename)
+
+## [0.2.0] - 2026-02-24
+
+- Added: A Mod installer to better handle poorly-packaged mods
+- Added: Button to download latest Malverk from GitHub (Texture Pack Manager - folder icon in Mods toolbar)
+- Fixed: Path Strings
+- Added: Buttons to open several useful files/folders/URLs (folder icon in Mods toolbar)
 
 ## [0.1.2] - 2025-05-28